fixes #2

+++ src/DistanceCalculator.java
@@ -1,6 +1,6 @@
 public class DistanceCalculator {
-    static double calculateDistance(double fuelVolume, double fuelConsumption){
-        double distance = (double) fuelVolume / fuelConsumption * 100;
+    static int calculateDistance(int fuelVolume, int fuelConsumption){
+        int distance = fuelVolume / fuelConsumption * 100;
         return distance;
     }
 }
